added getting rate endpoint - added scheduled logic - added cache

+++ src/main/resources/db/migrations/V1__creating_schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE exchange_rates
 (
     id                BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     base              VARCHAR(3)       NOT NULL,
-    quote_currency_id INTEGER          NOT NULL,
+    quote_currency_id BIGINT           NOT NULL,
     rate              DOUBLE PRECISION NOT NULL,
     CONSTRAINT fk_quote_currency FOREIGN KEY (quote_currency_id) REFERENCES currencies (id) ON DELETE CASCADE
 );
